Introducing more test about tourist spot,  and the logic to get passed the tests
In this case we add test about, get tourist spot by categories and regions. Also we  add test about to get all the tourist spots that exist. As well we add test to update and remove a tourist spot. And obviously we add the logic to get the test passed