Refactoring the test of Business Logic and Data Access because we mock and use Save() method of Repository
In this case we need to delete the line who mock the method Save() of the repository because, we introduce the function of that method into the operations who need to save changes in the context. The same things happens in DataAccess test, who use the method Save(), in this case we delete the line, because the method doesn´t exist.